update remaining stores to $state
