+++ src/lib/components/icons/index.ts
@@ -1,3 +1,5 @@
+export type IconProps = { size: number; fill?: string; class?: string };
+
 export { default as CollectionIcon } from "./collection.svelte";
 export { default as ZakuIcon } from "./zaku.svelte";
 export { default as FolderPlusIcon } from "./folder-plus.svelte";
