+++ src/lib/components/tree-item/tree-item-create.svelte
@@ -4,7 +4,7 @@
     import type { UnlistenFn } from "@tauri-apps/api/event";
 
     import { TREE_ITEM_TYPE } from "$lib/models";
-    import { createNewTreeItem, focussedTreeItem, zakuState } from "$lib/state.svelte";
+    import { zakuState, treeActionsState } from "$lib/state.svelte";
     import { cn, getMethodColorClass } from "$lib/utils/style";
     import { CollectionIcon } from "$lib/components/icons";
     import type { ValueOf } from "$lib/utils";
@@ -63,11 +63,11 @@
             inputName = "";
             await zakuState.synchronize();
 
-            focussedTreeItem.set({
+            treeActionsState.focussedItem = {
                 type: TREE_ITEM_TYPE.Collection,
                 parentRelativePath: createCollectionResult.value.parent_relative_path,
                 relativePath: createCollectionResult.value.relative_path,
-            });
+            };
 
             const createdCollection = document.querySelector(
                 `[data-current-path="${createCollectionResult.value.relative_path}"]`,
@@ -94,11 +94,11 @@
             inputName = "";
             await zakuState.synchronize();
 
-            focussedTreeItem.set({
+            treeActionsState.focussedItem = {
                 type: TREE_ITEM_TYPE.Request,
                 parentRelativePath: createRequestResult.value.parent_relative_path,
                 relativePath: createRequestResult.value.relative_path,
-            });
+            };
 
             const createdRequest = document.querySelector(
                 `[data-current-path="${createRequestResult.value.relative_path}"]`,
@@ -149,7 +149,7 @@
                 type="text"
                 onfocusout={async event => {
                     if (!isRelatedElementExcludedFromFocusOutTarget(event)) {
-                        createNewTreeItem.set(null);
+                        treeActionsState.createNewItem = null;
                         inputName = "";
                     } else {
                         inputName = "";
