+++ src/lib/components/tree-item/tree-item-preview.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    export let title: string;
+    let { title }: { title: string } = $props();
 </script>
 
 <div class="rounded-md bg-slate-600 px-2 py-0.5 text-sm text-white">
